// src/app/model/AddToCart.ts
export class addToCart{ 
 
 itemid:number=0;
  colorId: number=0
  quantity: number=0
  itemName: string="";
  itemDescription: string="";
  price: number=0;
  offerPrice: number=0;
  deliveryCharges: number=0;
  sizeName: string="";
  sizeId: number=0;
  colorName: string="";
  email: string|null="";
  count: number=0;
 
}